1. 2 Layers Of HashTable.（双层置换表） 2. Flex Size Of HashTable.（灵活设置置换表大小）

+++ ChineseChess/ChineseChess/Luna/Luna/Luna/Const/Luna+PreGenerate.h
@@ -198,7 +198,7 @@ LC_INLINE LCRowColumnMapState LCMoveMapGetColumnMapState(const LCRowColumn rank,
 }
 
 // MARK: - Zobrist 散列
-typedef UInt16 LCZobristHash;
+typedef UInt32 LCZobristHash;
 typedef UInt64 LCZobristKey;
 
 extern const LCChess LCZobristMap[LCChessLength];
